// packages/tldraw/src/types/TldrawEditorProps.ts
/**
 * Copyright (C) 2023 GIP-RECIA, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface TldrawEditorProps {
  debug?: boolean
  mode?: 'single' | 'multi'

  // Common properties
  persistanceApiUrl?: string
  assetsApiUrl?: string
  token?: string
  userInfoApiUrl?: string
  darkMode?: boolean
  readOnly?: boolean
  autoSave?: boolean
  autoSaveDelay?: number
  open?: boolean

  // Multi properties
  websocketApiUrl?: string
  roomId?: string
  initUrl?: string
  owner?: boolean
  clearOnLeave?: boolean
  leave?: boolean
}